requirements.txt
Flask==3.0.3
flask-cors==4.0.1
gspread==6.0.2
google-auth==2.30.0
google-cloud-speech==2.26.0